requirements.txt
django==5.0.3
django-crispy-forms==2.1
crispy-bootstrap4==2024.1
pillow==10.2.0
django-cities==0.6.1
django-multiselectfield==0.1.12